ss_d/requirements.txt
appdirs==1.4.3
crab==0.5.0
decorator==4.0.11
ez-setup==0.9
imageio==2.1.2
moviepy==0.2.2.13
music21==3.1.0
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
scikits.crab==0.1
scikits.learn==0.8.1
scipy==0.19.0
six==1.10.0
tqdm==4.11.2
unroll==0.1.0
